code shuffled after @adecler's comment
